new file:   assets/back-arrow-svgrepo-com.svg 	modified:   css/start.css 	modified:   index.html 	new file:   javascript/back.js 	modified:   javascript/bag.js 	modified:   javascript/locker.js 	modified:   javascript/main.js 	modified:   javascript/rules.js 	new file:   javascript/timer.js
