+++ index.html
@@ -13,6 +13,8 @@
     <script type="module" src="./javascript/locker.js"></script>
     <script type="module" src="./javascript/bag.js"></script>
     <script type="module" src="./javascript/rules.js"></script>
+    <script type="module" src="./javascript/timer.js"></script>
+    <script type="module" src="./javascript/back.js"></script>
 
 </head>
 
@@ -30,12 +32,12 @@
         </svg>
         <div class="timer-label">20:00</div>
     </div>
-
+    <input type="number" id="codeLocker">
     <a class="locker">.</a>
-    <a class="backLocker">.</a>
+    <a class="backLocker back"><img src="./assets/back-arrow-svgrepo-com.svg" alt="back-arrow"></a>
     <a class="bag">.</a>
-    <a class="backBag">.</a>
+    <a class="backBag back"><img src="./assets/back-arrow-svgrepo-com.svg" alt="back-arrow"></a>
     <a class="rules">.</a>
-    <a class="backRules">.</a>
+    <a class="backRules back"><img src="./assets/back-arrow-svgrepo-com.svg" alt="back-arrow"></a>
 </body>
 </html>
